requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
Authlib==1.6.4
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
fastapi==0.117.1
frozenlist==1.7.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langchain-weaviate==0.0.5
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langgraph-supervisor==0.0.29
langsmith==0.4.30
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
openai==1.109.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
postgrest==2.20.0
propcache==0.3.2
protobuf==6.32.1
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
realtime==2.20.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
simsimd==6.5.3
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
tavily-python==0.7.12
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
validators==0.35.0
weaviate-client==4.16.10
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
